Accept Docker progress on numbers >2GB
Update `ProgressUpdateEvent` to support images of a file size
>2GB without provoking build failures.

See gh-43328

+++ spring-boot-project/spring-boot-tools/spring-boot-buildpack-platform/src/main/java/org/springframework/boot/buildpack/platform/docker/ProgressUpdateEvent.java
@@ -22,6 +22,7 @@ import com.fasterxml.jackson.annotation.JsonCreator;
  * An {@link UpdateEvent} that includes progress information.
  *
  * @author Phillip Webb
+ * @author Wolfgang Kronberg
  * @since 2.3.0
  */
 public abstract class ProgressUpdateEvent extends UpdateEvent {
@@ -67,12 +68,12 @@ public abstract class ProgressUpdateEvent extends UpdateEvent {
 	 */
 	public static class ProgressDetail {
 
-		private final Integer current;
+		private final Long current;
 
-		private final Integer total;
+		private final Long total;
 
 		@JsonCreator
-		public ProgressDetail(Integer current, Integer total) {
+		public ProgressDetail(Long current, Long total) {
 			this.current = current;
 			this.total = total;
 		}
@@ -81,7 +82,7 @@ public abstract class ProgressUpdateEvent extends UpdateEvent {
 		 * Return the current progress value.
 		 * @return the current progress
 		 */
-		public int getCurrent() {
+		public long getCurrent() {
 			return this.current;
 		}
 
@@ -89,7 +90,7 @@ public abstract class ProgressUpdateEvent extends UpdateEvent {
 		 * Return the total progress possible value.
 		 * @return the total progress possible
 		 */
-		public int getTotal() {
+		public long getTotal() {
 			return this.total;
 		}
